Decompose Expensify.tsx: extract DelegateAccessHandler, FullstoryInitHandler, ScreenShareRequestModal
