+++ src/components/ScreenShareRequestModal.tsx
@@ -1,0 +1,32 @@
+import useLocalize from '@hooks/useLocalize';
+import useOnyx from '@hooks/useOnyx';
+import * as User from '@userActions/User';
+import ONYXKEYS from '@src/ONYXKEYS';
+import ConfirmModal from './ConfirmModal';
+
+/**
+ * Component that owns the SCREEN_SHARE_REQUEST Onyx subscription and renders
+ * the screen-share confirmation modal when a GuidesPlus agent requests it.
+ */
+function ScreenShareRequestModal() {
+    const {translate} = useLocalize();
+    const [screenShareRequest] = useOnyx(ONYXKEYS.SCREEN_SHARE_REQUEST);
+
+    if (!screenShareRequest) {
+        return null;
+    }
+
+    return (
+        <ConfirmModal
+            title={translate('guides.screenShare')}
+            onConfirm={() => User.joinScreenShare(screenShareRequest.accessToken, screenShareRequest.roomName)}
+            onCancel={User.clearScreenShareRequest}
+            prompt={translate('guides.screenShareRequest')}
+            confirmText={translate('common.join')}
+            cancelText={translate('common.decline')}
+            isVisible
+        />
+    );
+}
+
+export default ScreenShareRequestModal;
